crud mangnerger

+++ src/main/java/com/quynhtd/source_code_final/service/EmployeeService.java
@@ -1,31 +1,44 @@
 package com.quynhtd.source_code_final.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.quynhtd.source_code_final.entity.Employee;
-import com.quynhtd.source_code_final.respository.EmployeeRepository;
+import com.quynhtd.source_code_final.repository.EmployeeRepository;
 
 @Service
 public class EmployeeService {
-    @Autowired
-    private EmployeeRepository employeeRepository;
+	  @Autowired
+	    private EmployeeRepository employeeRepository;
 
-    public List<Employee> getAllEmployees() {
-        return employeeRepository.findAll();
-    }
+	    public List<Employee> getAllEmployees() {
+	        return employeeRepository.findAll();
+	    }
 
-    public Employee getEmployeeById(Long id) {
-        return employeeRepository.findById(id).orElse(null);
-    }
+	    public Optional<Employee> getEmployeeById(Long id) {
+	        return employeeRepository.findById(id);
+	    }
 
-    public Employee saveEmployee(Employee employee) {
-        return employeeRepository.save(employee);
-    }
+	    public Employee createEmployee(Employee Employee) {
+	        return employeeRepository.save(Employee);
+	    }
 
-    public void deleteEmployee(Long id) {
-        employeeRepository.deleteById(id);
-    }
-}
+	    public Employee updateEmployee(Long id, Employee Employee) {
+	        if (employeeRepository.existsById(id)) {
+	            Employee.setId(id);
+	            return employeeRepository.save(Employee);
+	        }
+	        return null;
+	    }
+
+	    public boolean deleteEmployee(Long id) {
+	        if (employeeRepository.existsById(id)) {
+	        	employeeRepository.deleteById(id);
+	            return true;
+	        }
+	        return false;
+	    }
+	}
